Allow the systemd services to restart endlessly
We saw that in our on prem deployments, network might be unreachable for a few
seconds when systemd services are starting up. With the current approach,
ssh-key-agent services would restart 5 times (default) and then remain dead.
Since we rely on that service to pull our keys in and allow us access to the
host, let's disable the default interval and allow the service to restart
forever, until network is up and it can start fine.