Support v prefix for tags
When binaries are added to the release via goreleaser - they reference
the version twice, once as a full tag, and second time stripping prefix
"v".

Add a Terraform local - that is the tag (version) with prefix "v"
stripped out.

Now we can tag / release version with prefix "v" and without.